+++ gusto_embedded/RELEASES.md
@@ -339,3 +339,13 @@ Based on:
 - [typescript v0.6.9] gusto_embedded
 ### Releases
 - [NPM v0.6.9] https://www.npmjs.com/package/@gusto/embedded-api/v/0.6.9 - gusto_embedded
+
+## 2025-09-02 21:08:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.607.0 (2.690.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.10] gusto_embedded
+### Releases
+- [NPM v0.6.10] https://www.npmjs.com/package/@gusto/embedded-api/v/0.6.10 - gusto_embedded